[egs] Gain 3x on dataset preprocessing time + factor out some functions
